application/requirements.txt
pika==1.3.2
pyserial==3.5
pydantic==1.10.0
python-dotenv==0.19.0